services/authz_service/pyproject.toml
[project]
name = "authz-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    "fastapi>=0.115.13",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.3",
]
